+++ myqueue/test/test_more.py
@@ -11,8 +11,8 @@ from ..task import task
 @pytest.mark.skipif(sys.version_info < (3, 9),
                     reason="requires Python 3.9 or higher")
 def test_completion():
-    from myqueue.utils import update_completion
-    update_completion(test=True)
+    from myqueue.utils import update_readme_and_completion
+    update_readme_and_completion(test=True)
 
 
 def test_api(mq):